feat(tutorial3): create GameBuilder
